backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
aiofiles
pydantic
pydantic-settings
biopython
psycopg[binary,pool]
redis
python-multipart
reportlab
